grad_project_backend-main(New)/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
face_recognition==1.3.0
opencv-python==4.10.0.84
numpy>=1.24
python-multipart==0.0.9
paho-mqtt==2.0.0
requests==2.31.0
